Edit bug

+++ app/Http/Controllers/CategoryController.php
@@ -59,9 +59,14 @@ class CategoryController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Category $category)
+    public function destroy($id)
     {
+        $category = Category::find($id);
+        if (!$category) {
+            return response()->json(['message' => 'Category not found'], 404);
+        }
         $category->delete();
+
         return response()->json(['message' => 'Category deleted successfully']);
     }
     /**
